ai_health_agent/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
langchain==0.2.14
langchain-openai==0.1.23
langgraph==0.2.34
chromadb==0.5.5
tiktoken==0.7.0
apscheduler==3.10.4
pandas==2.2.3
numpy==2.1.1
scipy==1.13.1
requests==2.32.3
streamlit==1.38.0
